Refactor sorting into a separate class
