+++ web-components/src/components/table/sorting.ts
@@ -1,0 +1,25 @@
+export type Order = 'asc' | 'desc';
+
+function desc<T>(a: T, b: T, orderBy: keyof T): any {
+  if (b[orderBy] < a[orderBy]) {
+    return -1;
+  }
+  if (b[orderBy] > a[orderBy]) {
+    return 1;
+  }
+  return 0;
+}
+
+export function stableSort<T>(array: ReadonlyArray<T>, cmp: (a: T, b: T) => number): any {
+  const stabilizedThis = array.map((el, index) => [el, index] as [T, number]);
+  stabilizedThis.sort((a, b) => {
+    const order = cmp(a[0], b[0]);
+    if (order !== 0) return order;
+    return a[1] - b[1];
+  });
+  return stabilizedThis.map(el => el[0]);
+}
+
+export function getSorting<T>(order: Order, orderBy: keyof T): (a: T, b: T) => number {
+  return order === 'desc' ? (a, b) => desc(a, b, orderBy) : (a, b) => -desc(a, b, orderBy);
+}
